for MPP-4020: add sp3_plans to backend and API
